Add missing type casting

+++ smsd/core.c
@@ -2120,7 +2120,7 @@ GSM_Error SMSD_MainLoop(GSM_SMSDConfig *Config, gboolean exit_on_failure, int ma
 
 		/* Sleep some time before another loop */
 		current_time = time(NULL);
-		lastsleep = round(difftime(current_time, lastloop));
+		lastsleep = (int)round(difftime(current_time, lastloop));
 		if (Config->loopsleep == 1) {
 			sleep(1);
 		} else if (lastsleep < Config->loopsleep) {
